Keep the launchd daemon alive
Working with multiple lima clusters in the recent weeks I found that
socket_vmnet is not running for unknown reason. The typical flow is
trying to start the clusters, and lima hostagent fails with connection
refused with /var/run/socket_vmnet. This happens to me one or more times
in the same day. Trying to run a stress test creating and destroying the
lima clusters 50 times fails after several runs and from the point of
the failure, all runs failed.

The issue seems to be that socket_vmnet is stopped by launched because
it seems to be idle and it is never started again. Adding the keep alive
option eliminated this issue.

With this change the daemon is kept running and it should restart after
failures.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>